recreated code and sorted files in folders
