+++ I2C_Scanner_V_1_1/main.py
@@ -2,542 +2,397 @@
 #also the ControlUnit(connected relays, Alarm state), Wiegand RFID card reader(all 4 pin addresses, card data, reader functionality)
 
 import machine
-from ili934xnew import ILI9341, color565
+from I2C_Scanner_V_1_1.display.ili934xnew import ILI9341, color565
 from machine import Pin, SPI, UART
-import m5stack
-from eeprom import EEPROM
-import glcdfont
+import I2C_Scanner_V_1_1.tools.m5stack as m5stack
+from I2C_Scanner_V_1_1.tools.eeprom import EEPROM
+import I2C_Scanner_V_1_1.display.glcdfont as glcdfont
 from micropython import const
-import tt14
-import tt24
+import I2C_Scanner_V_1_1.display.tt14 as tt14
+import I2C_Scanner_V_1_1.display.tt24 as tt24
 import random
-import tt32
+import I2C_Scanner_V_1_1.display.tt32 as tt32
 import os
 import utime
 import sys
-from MSG import MSG
-from Wiegand import Wiegand
-
-SCR_WIDTH = const(480)
-SCR_HEIGHT = const(320)
-SCR_ROT = const(3)
-CENTER_Y = int(SCR_WIDTH / 2)
-CENTER_X = int(SCR_HEIGHT / 2)
-
-TFT_CLK_PIN = const(18)
-TFT_MOSI_PIN = const(19)
-TFT_MISO_PIN = const(16)
-
-TFT_CS_PIN = const(17)
-TFT_RST_PIN = const(20)
-TFT_DC_PIN = const(21)
-
-Ttx=Pin(4)
-Trx=Pin(5)
-
-button = Pin(25, Pin.IN, Pin.PULL_UP)
-buzzer = Pin(24,Pin.OUT)
-results=False
-RFID_res=[]
-btn_count=0
-
-spi = SPI(#object used to configure the display of the LCD display
-    0,
-    baudrate=40000000,
-    miso=Pin(TFT_MISO_PIN),
-    mosi=Pin(TFT_MOSI_PIN),
-    sck=Pin(TFT_CLK_PIN)
-)
-
-sda=machine.Pin(0)
-scl=machine.Pin(1)
-i2c=machine.I2C(0,sda=sda,scl=scl,freq=400000)
-global atmega_found
-atmega_found=False
-true_found=False
-eeprom_check = False
-global mDevice#Global variable that says which Unit is used
-uartID=UART(1,baudrate=57600,timeout=1, invert=3)#Object used to find out if there is a ControlUnit connected or not
-
-M=MSG(print)
-M.Queue("GET","board_version")
-M.Send(True)
-M.Receive(True)
-checkU=uartID.read()#Variable used to find out if there is a ControlUnit connected or not
-print(checkU)
-
-display = ILI9341(#Object used to overall display the LCD display
-    spi,
-    cs=Pin(TFT_CS_PIN),
-    dc=Pin(TFT_DC_PIN),
-    rst=Pin(TFT_RST_PIN),
-    w=SCR_WIDTH,
-    h=SCR_HEIGHT,
-    r=SCR_ROT
-)
+from I2C_Scanner_V_1_1.tools.MSG import MSG
+from I2C_Scanner_V_1_1.tools.Wiegand import Wiegand
+from ComponentTests import ComponentTests
 
 
-eprom=EEPROM(#Object used to work with units EEPROM storage
-    addr=80,
-    pages=256,
-    bpp=64,
-    i2c=i2c,
-    at24x=0
-    )
+def newFrame():#Function for a new frame on the LCD display
+    device=CT.i2c.scan()
 
-try:
-    for i in range(10):
-        eprom.write(0, "test")
-        epromVal=eprom.read(0,4)
-except:
-    eeprom_check = False
-    print(eeprom_check)
-else:
-    eeprom_check = True
-    print(eeprom_check)
-eprom.wipe()
-    
-def RandomString(length=6):
-    characters = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'
-    random_string = ''.join(random.choice(characters) for _ in range(length))
-    return random_string
+    CT.display.erase()
+    CT.display.fill_rectangle(0 , 0, 240, 320, 65535)
+    CT.display.fill_rectangle(0 , 0, 240, 50, 008000)
+    CT.display.fill_rectangle(0 , 50, 240, 5, 000033)
 
-    
-#2D array that has the I2C addresses, hardware name and its status in the scanner(if there are more hardware connected, than add its info in the array)
-i2c_devices=[['104','EEPROM','False'],['80','RTC','False']]
-statusArr=[]
+    if check_uart is not None and device != [] and check_uart != b'\x00':#Checks if there is a ControlUnit connected
+        CT.display.set_pos(120, 285)
+        CT.display.set_font(tt24)
+        CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
 
+        CT.display.write("ControlUnit")
 
-def EEPROM_check(y,i2c_devices,results):#EEPROM testing function
-    global statusArr
-    newFrame(2)
-    random_arr = []
-    display.set_pos(10, 60)
-    display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-    display.print('Testing EEPROM')
-    try:
+    elif check_uart==b'\x00' and device != []:#Otherwise it shows that there is a TTunit connected
+        CT.display.set_pos(165, 285)
+        CT.display.set_font(tt24)
+        CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
 
-        for i in range(256):
-            random_arr.append(RandomString())
-            eprom.write(i*64, random_arr[i])
-        
-        statusArr.append('* EEPROM communication: Write OK')
+        CT.display.write("TTunit")
 
-    except:#Finds out if the EEPROM writing works
-        statusArr.append('* EEPROM communication: Write ERR')
-        i2c_devices[1][2]='False'
-
-    try:
-        
-        for i in range(256):
-            if random_arr[i] not in eprom.read(i*64,6):
-                break
-        if i == 255:       
-            statusArr.append('* EEPROM communication: Read OK')
-        else:
-            statusArr.append('* EEPROM communication: Read ERR')
-
-    except:#Finds out if the EEPROM reading works
-        statusArr.append('* EEPROM communication: Read ERR')
-        i2c_devices[1][2]='False'
-
-    eprom.wipe()
-    
-def RFID_check(): # RFID card reader testing function
-    global results, RFID_res
-    RFID_adr=[[2,3,6,8],[9,10,7,11],[14,15,26,13],[28,29,6,27]]#RFID card reader addresses
-    RFID_res=[]
-    pg=9
-    i=0
-    loop=0
-    for add in RFID_adr:#Goes through each pin address to test
-        pg+=1
-        i+=1
-        led=Pin(add[2],Pin.OUT)
-        buz=Pin(add[3],Pin.OUT)
-        wiegand_reader = Wiegand(add[0], add[1])#Sets the pin addresses to work with the wiegand protocol
-        newFrame(pg)
-        display.set_pos(170, 10)
-        display.set_font(tt14)
-        display.write("PRESS---->")
-        display.set_pos(10, 60)
-        display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-        display.print('* '+str(i)+'. Wiegand device address testing')
-        display.fill_rectangle(5, 90, 215, 45, 008000)
-        display.set_pos(10, 100)
-        display.set_font(tt24)
-        display.print("For next test press")
-        display.set_font(tt14)
-        display.set_pos(10, 160)
-        display.print("Wiegand Card Data:")
-        display.set_pos(10, 190)
-        display.print("Wiegand Card Data(rev):")
-        display.set_pos(10, 220)
-        display.print("Wiegand Type(bits):")
-
-        while True:
-            if wiegand_reader.available() and loop==0:#Checks if card has been scanned by the RFID card reader
-                for x in range(2):
-                    led.high()
-                    buz.high()
-                    utime.sleep(0.4)
-                    led.low()
-                    buz.low()
-                    utime.sleep(0.4)
-                if loop==0:
-                    loop=1
-                    card_code = wiegand_reader.GetCode()# Gets the card code
-                    card_revCode=wiegand_reader.GetRevCode()
-                    card_type = wiegand_reader.GetType()# Gets the RFID bit type
-                    display.set_pos(10, 160)
-                    display.print("RFID Card Data:" + str(card_code))
-                    display.set_pos(10, 190)
-                    display.print("Wiegand Card Data(rev):" + str(card_revCode))
-                    display.set_pos(10, 220)
-                    display.print("Wiegand Type(bits):" + str(card_type))
-                    RFID_res.append(str(i)+'.')#Puts the RFID card reader functionality statuss
-                    if results == True:  # Used for result identification further on in the code
-                        results = False
-                    else:
-                        results = True
-            button_pressed = button.value() == 0  # Check if the button is pressed
-
-            if button_pressed:
-                loop=0
-                break
-    if RFID_res==[]:#If there aren't any working card readers it puts in the status array information for results showcase
-        statusArr.append('* 0 Wiegand readers work: ERR')
-    elif RFID_res != []:#If there are working card readers it puts in the status array information for results showcase
-        ind=str(RFID_res)
-        statusArr.append('* '+ind+' Wiegand readers work')
-
-def ATmega_relay_check():#Relay testing function
-    signalArr=["turnstile1_a","turnstile1_b","turnstile2_a","turnstile2_b","button1","button2"]#Array of all test-needed relays
-    pg=4
-    for relayName in signalArr:#Goes through each test-needed relays
-        newFrame(pg)
-        display.set_pos(10, 60)
-        display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-        display.print('* '+relayName+' testing')
-        display.fill_rectangle(5, 90, 180, 45, 008000)
-        display.set_pos(10, 100)
-        display.set_font(tt24)
-        display.print("Listen for relay!")
-        display.set_font(tt14)
-        M.Queue("SIGNAL",relayName)#Signals ControlUnit to turn on relay
-        M.Send(True)
-        pg=pg+1
-        utime.sleep(3)
-        
-def ATmega_check(y,results):#ControlUnit testing function
-    global statusArr
-    global atmega_found
-    newFrame(3)
-    display.set_pos(10, 60)
-    display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-    display.print('Testing ControlUnit')
-    atmega_found=False#Variable used for the ControlUnits functionality identification
-    ATmega=M.HWCheck()
-    y+=30
-    if ATmega ==True:
-        atmega_found=True#Sets the ConrolUnits status as True
-        statusArr.append('* ConUnit communication: Write OK')
-        statusArr.append('* ConUnit communication: Read OK')
-        if results==True:#Used for result identification further on in the code
-                results=False
-        else:
-                results=True
-    else:#If the hardware check function gives out False value
-        statusArr.append('* ConUnit communication: Write ERR')
-        statusArr.append('* ConUnit communication: Read ERR')
-    display.set_pos(130, 10)
-    display.set_font(tt14)
-    display.write("TEST RELAYS---->")
-    while True:
-        button_pressed = button.value() == 0  # Check if the button is pressed
-        if button_pressed:
-            ATmega_relay_check()
-            RFID_check()
-            break
-        
-rtc=machine.RTC()#Real-Time-Clock defining
-print(rtc.datetime())
-
-def RTC_check(y,i2c_devices,results):#Real-Time-Clock hardware testing function
-    global statusArr
-    newFrame(1)
-    display.set_pos(10, 60)
-    display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-    display.print('Testing RTC')
-    oldRtc=rtc.datetime()#Variable used to set RTC time to currect time
-    try:
-        rtc.datetime((2020, 1, 21, 2, 10, 32, 36, 0))#Writes RTC time
-    except:#Finds out if the RTC writing works
-        statusArr.append('* RTC communication: Write ERR')
-        i2c_devices[1][2]='False'
-    else:
-         rtc.datetime((2020, 1, 21, 2, 10, 32, 36, 0))#Writes the RTC time
-    try:
-        print(rtc.datetime())#Reads the RTC time
-    except:#Finds out if RTC hardware even works
-        statusArr.append('* RTC communication: Read ERR')
-        i2c_devices[1][2]='False'
-    try:
-         if rtc.datetime()[0] == 2020 and rtc.datetime()[4] == 10:#Checks if the RTC reads and writes perfectly
-             display.set_pos(10, y)
-    except:#If RTC doesn't work then it writes an error in RTC communication
-        statusArr.append('* RTC communication ERR')
-        i2c_devices[1][2]='False'
-    else:
-        if rtc.datetime()[0] == 2020 and rtc.datetime()[4] == 10:#Checks if the RTC reads and writes perfectly
-            statusArr.append('* RTC communication: Write OK')
-            statusArr.append('* RTC communication: Read OK')
-            if results==True:#Used for result identification further on in the code
-                results=False
-            else:
-                results=True
-        else:#If RTC doesn't work then it writes an error in RTC communication
-            statusArr.append('* RTC communication ERR')
-            i2c_devices[1][2]='False'
-    rtc.datetime(oldRtc)
-
-def newFrame(pg):#Function for a new frame on the LCD display
-    device=i2c.scan()
-    print("devices")
-    print(device)
-    display.erase()
-    display.fill_rectangle(0 , 0, 240, 320, 65535)
-    display.fill_rectangle(0 , 0, 240, 50, 008000)
-    display.fill_rectangle(0 , 50, 240, 5, 000033)
-    if checkU is not None and device != [] and checkU != b'\x00':#Checks if there is a ControlUnit connected
-        display.set_pos(120, 285)
-        display.set_font(tt24)
-        display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-        display.write("ControlUnit")
-    elif checkU==b'\x00' and device != []:#Otherwise it shows that there is a TTunit connected
-        display.set_pos(165, 285)
-        display.set_font(tt24)
-        display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-        display.write("TTunit")
     elif device == []:#If there aren't any devices connected to RPi
-        display.set_pos(130, 285)
-        display.set_font(tt24)
-        display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-        display.write("No device")
-    display.fill_rectangle(10 , 16, 85, 30, 111111)
-    display.fill_rectangle(5 , 11, 85, 30, 65535)
-    display.set_font(tt24)
-    display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-    display.set_pos(10, 15)
-    display.write('IN')
-    display.set_color(color565(0, 255, 0), color565(255, 255, 255))
-    display.set_pos(30, 15)
-    display.write('PASS')
-    display.set_font(tt14)
-    display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-    if pg != 0 and pg != "res":#If there are any test numbers set it sets to needed test number
-        display.set_pos(10, 300)
-        display.write(str(pg)+". test")
-    elif pg == "res":#If there is given a result signal it will set it as Results on frame
-        display.set_pos(10, 300)
-        display.write("Results")
-    display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+        CT.display.set_pos(130, 285)
+        CT.display.set_font(tt24)
+        CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+
+        CT.display.write("No device")
+
+    CT.display.fill_rectangle(10 , 16, 85, 30, 111111)
+    CT.display.fill_rectangle(5 , 11, 85, 30, 65535)
+
+    CT.display.set_font(tt24)
+    CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+    CT.display.set_pos(10, 15)
+    CT.display.write('IN')
+
+    CT.display.set_color(color565(0, 255, 0), color565(255, 255, 255))
+    CT.display.set_pos(30, 15)
+    CT.display.write('PASS')
+
+    CT.display.set_font(tt14)
+    CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+
+    if CT.page != 0 and CT.page != "res":#If there are any test numbers set it sets to needed test number
+        CT.display.set_pos(10, 300)
+        CT.display.write(str(CT.page)+". test")
+
+    elif CT.page == "res":#If there is given a result signal it will set it as Results on frame
+        CT.display.set_pos(10, 300)
+        CT.display.write("Results")
+
+CT = ComponentTests(newFrame)
+
+check_uart = CT.UartComPreTest()
+check_eeprom = CT.EepromPreTest()
+
+def TestSuccess():
+    if CT.used_device == "ATmega":
+
+        if CT.i2c_devices[0][2]=="True" and CT.i2c_devices[1][2]=="True":
+
+            if '* ConUnit communication: Write OK' in CT.results_list  and CT.rfid_res==['1.','2.','3.','4.']:
+                return True
+            
+        return False
+
+    elif CT.used_device == "I2C":
+
+        if CT.i2c_devices[0][2]=="True" and CT.i2c_devices[1][2]=="True":
+            return True
+        
+        return False
+
     
 def scan_i2c():#The main scanner function
-    global mDevice, RFID_res, statusArr
-    devices = i2c.scan()
+    devices = CT.i2c.scan()
     x=0
-    rtc_found = False
-    y=60
 
     if devices:#If there are I2C hardware used
         newFrame(0)
-        display.set_pos(10, 60)
-        display.print('* I2C devices found:')
+        CT.display.set_pos(10, 60)
+        CT.display.print('* I2C devices found:')
+
         print("Found I2C devices")
-        display.set_pos(135, 60)
-        lvl=30
+
         col = 60
+
         for device in devices:
-            display.set_pos(135, col)
-            display.print(f"{hex(device)};")
+
+            CT.display.set_pos(135, col)
+            CT.display.print(f"{hex(device)};")
+
             col = col+15
-        if checkU is not None and checkU!=b'\x00':#Checks if there is a ControlUnit connected
+
+        if check_uart is not None and check_uart!=b'\x00':#Checks if there is a ControlUnit connected
+
+            CT.used_device="ATmega"#Names the Unit according to the test
+
             col = col+10
-            display.set_pos(10, col)
-            display.print('* ControlUnit device found')
+
+            CT.display.set_pos(10, col)
+            CT.display.print('* ControlUnit device found')
+
             print("Found ControlUnit device")
+
+        elif check_uart==b'\x00':
+            CT.used_device="I2C"#Names the Unit according to the test
+
+            col = col+10
+
+            CT.display.set_pos(10, col)
+            CT.display.print('* TTunit device found')
+
+            print("Found TTunit device")
+
         utime.sleep(2)
+
         newFrame(0)
+        col = 30
+
         for device in devices:#Makes the I2C hardware status true for further results identification
             x=0
-            for hexs, name, status in i2c_devices:
-             if hexs == str(device):
-                lvl=lvl+30
-                display.set_pos(10, lvl)
-                display.print('* '+name+' connected')
-                i2c_devices[x][2]='True'
-             x+=1
-        if checkU is not None and checkU!=b'\x00':#Checks if there is a ControlUnit connected
-            mDevice="ATmega"#Names the Unit according to the test
-            lvl=lvl+30
-            display.set_pos(10, lvl)
-            display.print('* ControlUnit connected')
-        elif checkU==b'\x00':
-            mDevice="I2C"#Names the Unit according to the test
+            for hexs, name, status in CT.i2c_devices:
+
+                if hexs == str(device):
+                    col+=30
+
+                    CT.display.set_pos(10, col)
+                    CT.display.print('* '+name+' connected')
+
+                    CT.i2c_devices[x][2]='True'
+
+                x+=1
+
+        if check_uart is not None and check_uart!=b'\x00':#Checks if there is a ControlUnit connected
+
+            col+=30
+            CT.display.set_pos(10, col)
+            CT.display.print('* ControlUnit connected')
+
         print(str(x)+" I2C devices were found")
+
         utime.sleep(1)
-        for device in i2c_devices:
+
+        for device in CT.i2c_devices:
             if 'True' in device:#Checks if the device has a "True" value
-                true_found = True
+                i2c_found = True
+
                 break
-        if true_found:#If there is at least one "True"    
-                        y=60
-                        RTC_check(60,i2c_devices,results)#Runs the RTC_check function, y is used for the storing of the display location
-                        EEPROM_check(y,i2c_devices,results)#Runs the EEPROM_check function, y is used for the storing of the display location
-                        if checkU is not None and checkU!=b'\x00':#Checks if there is a ControlUnit connected
-                            ATmega_check(y,results)#Runs the ATmega_check function, y is used for the storing of the display location
+        if i2c_found:#If there is at least one "True"  
+                          
+                        CT.page = 1
+                        CT.RTC_check()#Runs the RTC_check function, y is used for the storing of the display location
+
+                        CT.page=2
+                        CT.EEPROM_check()#Runs the EEPROM_check function, y is used for the storing of the display location
+
+                        if check_uart is not None and check_uart!=b'\x00':#Checks if there is a ControlUnit connected
+
+                            CT.page=3
+                            CT.ATmega_check()#Runs the ATmega_check function, y is used for the storing of the display location
+
+                            while True:
+                                button_pressed = CT.button.value() == 0  # Check if the button is pressed
+                                if button_pressed:
+
+                                    CT.page = 4
+                                    CT.ATmega_relay_check()
+
+                                    CT.page = 9
+                                    CT.RFID_check()
+
+                                    break
+
+
                         utime.sleep(0.3)
-                        if checkU is not None and checkU!=b'\x00':
-                            display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-                            display.set_pos(170, 10)
-                            display.set_font(tt14)
-                            display.write("ON---------->")
-                            display.set_pos(10, 265)
-                            display.print('Alarm to confirm scan results')
-                        elif checkU==b'\x00' and device != []:
-                            display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-                            display.set_pos(170, 10)
-                            display.set_font(tt14)
-                            display.write("Buzz------>")
-                            display.set_pos(10, 265)
-                            display.print('Buzz to confirm scan results')
-                    
+                        if check_uart is not None and check_uart!=b'\x00':
+
+                            CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+                            CT.display.set_pos(170, 10)
+                            CT.display.set_font(tt14)
+                            CT.display.write("ON---------->")
+
+                            CT.display.set_pos(10, 265)
+                            CT.display.print('Alarm to confirm scan results')
+
+                        elif check_uart==b'\x00' and device != []:
+
+                            CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+                            CT.display.set_pos(170, 10)
+                            CT.display.set_font(tt14)
+                            CT.display.write("Buzz------>")
+
+                            CT.display.set_pos(10, 265)
+                            CT.display.print('Buzz to confirm scan results')
+
         else:#If there isn't any "True" found
-          t=60
+          
+          col = 60
           newFrame(0)
-          for hexs, names, status in i2c_devices:
+
+          for hexs, names, status in CT.i2c_devices:
             if status=="False":#Checks the I2C hardware for "False" values
-                display.set_pos(0, t)
-                display.set_color(color565(0, 0, 255), color565(255, 255, 255))
-                display.print('* '+names+' not found.')
+
+                CT.display.set_pos(0, col)
+                CT.display.set_color(color565(0, 0, 255), color565(255, 255, 255))
+                CT.display.print('* '+names+' not found.')
+
                 print("The scanner didn't find "+names+" device")
-                t+=30
+
+                col+=30
            
     else:#If there aren't any I2C hardware found
         newFrame(0)
+
         print(devices)
-        display.set_pos(10, 60)
-        display.set_color(color565(0, 0, 255), color565(255, 255, 255))
-        display.print('* I2C devices not found.')
+
+        CT.display.set_pos(10, 60)
+        CT.display.set_color(color565(0, 0, 255), color565(255, 255, 255))
+        CT.display.print('* I2C devices not found.')
+
         print("There weren't any I2C devices found")
-        if eeprom_check == False:
-            display.set_pos(10, 90)
-            display.print('* Likely a problem with EEPROM.')
-        mDevice=None
-        btn_count=0
-        
-def scannerRestart(btn_count):#Function that lets the user restart scanner functions
-   global mDevice, RFID_res
-   if mDevice!=None:
-    while True:#Meant for pressing a button
-     if button.value() == 0:
-      for i in range(2):
-        btn_count += 1
-        if btn_count ==1:#If the button is pressed one time goes through a buzzer test
-            newFrame(0)
-            if checkU is not None and checkU!=b'\x00':#If a ControlUnit is connected
-                display.set_pos(10, 60)
-                display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-                display.print("Alarming")
-                print("Buzzing!")
-                M.Queue("SET","alarm_on_state",value=1)
-                M.Send(True)
-                utime.sleep(3)
-                M.Queue("SET","alarm_on_state",value=0)
-                M.Send(True)
-            elif checkU==b'\x00':#If a TTunit is connected
-                display.set_pos(10, 60)
-                display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-                display.print("Buzzing")
-                display.set_pos(170, 10)
-                display.set_font(tt14)
-                print("Buzzing!")
-                for x in range(5):
-                    buzzer.on()
-                    utime.sleep(0.1)
-                    buzzer.off()
-                    utime.sleep(0.1)
-        elif btn_count==2:#If button is pressed two times shows results and lets the user restart the main scanner function
-            newFrame("res")
-            display.set_pos(160, 10)
-            display.set_font(tt14)
-            display.write("RESTART---->")
-            if (i2c_devices[0][2]=="True" and i2c_devices[1][2]=="True" and atmega_found==True and mDevice=="ATmega" and RFID_res==['1.','2.','3.','4.']) or (i2c_devices[0][2]=="True" and i2c_devices[1][2]=="True" and mDevice=="I2C"):
-                #Checks the status of all of the hardware, if every hardware is found and tested successfully
-                k=60
-                i=0
-                for stat in statusArr:
-                    i=i+1
-                    display.set_pos(5, k)
-                    display.set_font(tt14)
-                    if "ERR" in stat or "ERROR" in stat:
-                        display.set_color(color565(0, 0, 255), color565(255, 255, 255))
-                        print("err")
-                    display.write(stat)
-                    if (i < 7 and mDevice=="ATmega") or mDevice=="I2C":
-                        display.fill_rectangle(20 , k+15, 200, 2, 000033)
-                    k=k+25
-                    display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-                display.fill_rectangle(0 , 240, 240, 5, 008000)
-                display.set_pos(30, 250)
-                display.set_font(tt24)
-                display.set_color(color565(0, 255, 0), color565(255, 255, 255))
-                display.write("Test successful!")
-                print("The I2C scanner test was successful!")
-            else:#If the hardware status isn't successful
-                k=60
-                i=0
-                for stat in statusArr:
-                    i=i+1
-                    display.set_pos(5, k)
-                    display.set_font(tt14)
-                    if RFID_res != ['1.','2.','3.','4.'] and "readers" in stat:
-                        display.set_color(color565(0, 0, 255), color565(255, 255, 255))
-                    if ("ERR" in stat or "ERROR" in stat):
-                        display.set_color(color565(0, 0, 255), color565(255, 255, 255))
-                    display.write(stat)
-                    if (i < 7 and mDevice=="ATmega") or mDevice=="I2C":
-                        display.fill_rectangle(20 , k+15, 200, 2, 000033)
-                    k=k+25
-                    display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-                display.set_pos(30, 250)
-                display.fill_rectangle(0 , 240, 240, 5, 000030)
-                display.set_font(tt24)
-                display.set_color(color565(0, 0, 255), color565(255, 255, 255))
-                display.write("Test unsuccessful!")
-                print("The I2C scanner test was unsuccessful!")
-            return btn_count
-   elif mDevice==None:#If there isn't any device connected
-       btn_count=0
-       return btn_count
+
+        if check_eeprom == False:
+            CT.display.set_pos(10, 90)
+            CT.display.print('* Likely a problem with EEPROM.')
+
+        CT.used_device = None
+
+def TestResults():
+    if CT.used_device != None:
+
+        while True:
+            if CT.main_button.value() == 0:
+
+                newFrame(0)
+
+                if check_uart is not None and check_uart!=b'\x00':#If a ControlUnit is connected
+
+                    CT.display.set_pos(10, 60)
+                    CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+                    CT.display.print("Alarming")
+
+                    print("Buzzing!")
+
+                    CT.M.Queue("SET","alarm_on_state",value=1)
+                    CT.M.Send(True)
+
+                    utime.sleep(3)
+
+                    CT.M.Queue("SET","alarm_on_state",value=0)
+                    CT.M.Send(True)
+
+                elif check_uart==b'\x00':#If a TTunit is connected
+
+                    CT.display.set_pos(10, 60)
+                    CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+                    CT.display.print("Buzzing")
+
+                    CT.display.set_pos(170, 10)
+                    CT.display.set_font(tt14)
+
+                    print("Buzzing!")
+
+                    for x in range(5):
+                        CT.TT_buzzer.on()
+                        utime.sleep(0.1)
+
+                        CT.TT_buzzer.off()
+                        utime.sleep(0.1)
+            elif CT.main_button.value() == 1:
+
+                break    
+
+        while True:
+            if CT.main_button.value() == 0:      
+
+                newFrame("res")
+
+                CT.display.set_pos(160, 10)
+                CT.display.set_font(tt14)
+                CT.display.write("RESTART---->")
+
+                if TestSuccess():
+                    #Checks the status of all of the hardware, if every hardware is found and tested successfully
+                    col = 60
+
+                    for i, stat in enumerate(CT.results_list):
+                        i+=1
+
+                        CT.display.set_pos(5, y)
+                        CT.display.set_font(tt14)
+
+                        if "ERR" in stat or "ERROR" in stat:
+                            CT.display.set_color(color565(0, 0, 255), color565(255, 255, 255))
+
+                            print("err")
+
+                        CT.display.write(stat)
+
+                        if (i < 7 and CT.used_device=="ATmega") or CT.used_device=="I2C":
+
+                            CT.display.fill_rectangle(20 , col+15, 200, 2, 000033)
+
+                        col+=25
+
+                        CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+
+                    CT.display.fill_rectangle(0 , 240, 240, 5, 008000)
+
+                    CT.display.set_pos(30, 250)
+                    CT.display.set_font(tt24)
+                    CT.display.set_color(color565(0, 255, 0), color565(255, 255, 255))
+                    CT.display.write("Test successful!")
+
+                    print("The I2C scanner test was successful!")
+                else:#If the hardware status isn't successful
+
+                    col=60
+
+                    for i, stat in enumerate(CT.results_list):
+                        i=i+1
+
+                        CT.display.set_pos(5, col)
+                        CT.display.set_font(tt14)
+
+                        if CT.rfid_res != ['1.','2.','3.','4.'] and "readers" in stat:
+
+                            CT.display.set_color(color565(0, 0, 255), color565(255, 255, 255))
+
+                        elif ("ERR" in stat or "ERROR" in stat):
+
+                            CT.display.set_color(color565(0, 0, 255), color565(255, 255, 255))
+
+                        CT.display.write(stat)
+
+                        if (i < 7 and CT.used_device=="ATmega") or CT.used_device=="I2C":
+
+                            CT.display.fill_rectangle(20 , col+15, 200, 2, 000033)
+
+                        col+=25
+
+                        CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+
+                    CT.display.fill_rectangle(0 , 240, 240, 5, 000030)
+
+                    CT.display.set_pos(30, 250)
+                    CT.display.set_font(tt24)
+                    CT.display.set_color(color565(0, 0, 255), color565(255, 255, 255))
+                    CT.display.write("Test unsuccessful!")
+
+                    print("The I2C scanner test was unsuccessful!")
     
 newFrame(0)#The first starting frame
+
 print("Starting program")
-display.set_pos(10, 60)
-display.set_color(color565(0, 0, 0), color565(255, 255, 255))
-display.print("* To Start press button")
-display.set_pos(170, 10)
-display.set_font(tt14)
-display.write("PRESS---->")
+
+CT.display.set_pos(10, 60)
+CT.display.set_color(color565(0, 0, 0), color565(255, 255, 255))
+CT.display.print("* To Start press button")
+
+CT.display.set_pos(170, 10)
+CT.display.set_font(tt14)
+CT.display.write("PRESS---->")
+
 while True:#If button is pressed restarts the main scanner function
-    if button.value() == 0:
-        if btn_count ==0:#If the start/restart button hasn't been pressed or has been reset
-            print("1st")
-            scan_i2c()
-            btn_count=scannerRestart(btn_count)
-        elif btn_count==2:#If the start/restart button has been pressed just once it confirms the test and restarts
-            print("2nd")
-            btn_count=0
-            statusArr=[]
-            RFID_res=[]
+    if CT.main_button.value() == 1:
+        print("Test starting")
+
+        CT.ResetForTest()
+
+        scan_i2c()
+
+        TestResults()
+
